Created, implemented custom exceptions and Factories use enums to create obstacles.
Created a namespace CustomExceptions, where it has 3 Classes all inherited ArgumentException: DoubleArgumentException, IntArgumentException, and StringArgumentException. Most of the classes have utilised throwing these exceptions BUT none of them are handled yet. So in this commit, the program can't be ran just yet. The error message aren't part of these classes, however individual messages are passed on to the exceptions depending on the use case.

Factories are now simplified into one factory ObstacleFactory. This now has enums which corresponds to the type of obstacle. A switch statement is implemented in ObstacleFactory.AddToGrid which handles all types of obstacles.