Created a static Didutron.ArgsCount class.
This class has all of the argument count constants and a helper function to compare a string[] args with an int expectedCount. Which will throw an IncorrectNumberOfArgumentsException when the length of args do not match with the expected count. All methods that needs argument count checking have implemented this change.